Move coref scoring code to scorer.py
Includes some renames to make names less generic.
